requirements.txt
    djangorestframework
    cloudinary
    pillow
    python-decouple
    psycopg2-binary